+++ src/main/java/com/github/mullerdaniil/miranda/module/progressTracking/event/TrackDataUpdatedEvent.java
@@ -1,0 +1,5 @@
+package com.github.mullerdaniil.miranda.module.progressTracking.event;
+
+public record TrackDataUpdatedEvent(Integer trackId,
+                                    boolean updateAll) {
+}