+++ src/main/java/com/github/mullerdaniil/miranda/module/progressTracking/exception/ProgressPointServiceException.java
@@ -1,0 +1,7 @@
+package com.github.mullerdaniil.miranda.module.progressTracking.exception;
+
+public class ProgressPointServiceException extends RuntimeException {
+    public ProgressPointServiceException(String message) {
+        super(message);
+    }
+}