+++ src/main/java/com/github/mullerdaniil/miranda/module/progressTracking/exception/TrackServiceException.java
@@ -1,0 +1,7 @@
+package com.github.mullerdaniil.miranda.module.progressTracking.exception;
+
+public class TrackServiceException extends RuntimeException {
+    public TrackServiceException(String message) {
+        super(message);
+    }
+}